feat: Changed RefId to int

+++ 0_Framework/Application/ZarinPal/VerificationResponse.cs
@@ -3,6 +3,6 @@
     public class VerificationResponse
     {
         public int Status { get; set; }
-        public long RefID { get; set; }
+        public int RefID { get; set; }
     }
 }
